マップの修正
