+++ Assets/Scripts/Game/Note.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 244572ee2be3bc844857a987a50c7280
+guid: 3f51eb9894e409243a011eab13d4a9ff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
